+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Version Changes of AngloChan
 
+## [0.4.0]
+
+-   Added API Routes for Boards and Threads
+-   Added Controllers for Boards and Threads
+-   Added URL catching to all resolve to Svelte application
+
 ## [0.3.0]
 
 -   Modified migration models to be nullable